# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/categories": {
      "post": {
        "description": "Cria uma nova categoria de tarefas para o usuário autenticado.",
        "operationId": "CategoriesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Categoria criada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryDto"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "409": {
            "description": "Nome de categoria já utilizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Criar nova categoria",
        "tags": [
          "categories"
        ]
      },
      "get": {
        "description": "Lista todas as categorias ativas do usuário autenticado, com a contagem de tarefas.",
        "operationId": "CategoriesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lista de categorias",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CategoryDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar categorias do usuário",
        "tags": [
          "categories"
        ]
      }
    },
    "/api/v1/categories/{id}": {
      "put": {
        "description": "Atualiza o nome ou a cor de uma categoria pertencente ao usuário.",
        "operationId": "CategoriesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Categoria atualizada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryDto"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "403": {
            "description": "Acesso negado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "404": {
            "description": "Categoria não encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "409": {
            "description": "Nome de categoria já utilizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Atualizar categoria",
        "tags": [
          "categories"
        ]
      },
      "delete": {
        "description": "Realiza a remoção lógica da categoria e desvincula as tarefas associadas.",
        "operationId": "CategoriesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Categoria excluída com sucesso"
          },
          "401": {
            "description": "Não autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "403": {
            "description": "Acesso negado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "404": {
            "description": "Categoria não encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Excluir categoria (remoção lógica)",
        "tags": [
          "categories"
        ]
      }
    }
  },
  "info": {
    "title": "AppStart API",
    "description": "API REST do AppStart - Especificação OpenAPI 3.0",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "appstart_session",
        "description": "Cookie de sessão local opaca"
      }
    },
    "schemas": {
      "ProblemDetailsDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "https://appstart.example/problems/validation-error"
          },
          "title": {
            "type": "string",
            "example": "Dados inválidos"
          },
          "status": {
            "type": "number",
            "example": 400
          },
          "detail": {
            "type": "string",
            "example": "Um ou mais campos precisam ser corrigidos."
          },
          "instance": {
            "type": "string",
            "example": "/api/v1/users"
          },
          "code": {
            "type": "string",
            "example": "VALIDATION_ERROR"
          },
          "requestId": {
            "type": "string",
            "example": "01J..."
          },
          "timestamp": {
            "type": "string",
            "example": "2026-08-31T20:00:00.000Z"
          },
          "errors": {
            "type": "object",
            "example": {
              "email": [
                "Informe um e-mail válido."
              ]
            },
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "code",
          "timestamp"
        ]
      },
...
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nome da categoria",
            "example": "Trabalho",
            "minLength": 1,
            "maxLength": 50
          },
          "color": {
            "type": "string",
            "description": "Cor em hexadecimal",
            "example": "#3B82F6",
            "default": "#3B82F6"
          }
        },
        "required": [
          "name"
        ]
      },
      "CategoryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Identificador único da categoria",
            "example": "c0eebc99-9c0b-4ef8-bb6d-6bb9bd380a33"
          },
          "name": {
            "type": "string",
            "description": "Nome da categoria",
            "example": "Trabalho"
          },
          "color": {
            "type": "string",
            "description": "Cor em hexadecimal",
            "example": "#3B82F6"
          },
          "taskCount": {
            "type": "number",
            "description": "Quantidade de tarefas ativas associadas",
            "example": 4
          },
          "createdAt": {
            "type": "string",
            "description": "Data de criação"
          },
          "updatedAt": {
            "type": "string",
            "description": "Data de última atualização"
          }
        },
        "required": [
          "id",
          "name",
          "color",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nome da categoria",
            "minLength": 1,
            "maxLength": 50
          },
          "color": {
            "type": "string",
            "description": "Cor em hexadecimal",
            "example": "#3B82F6"
          }
        }
      }
    }
  }
}